+++ BFKit/Additions/Foundation/NSString+BFKit.m
@@ -32,6 +32,7 @@
 + (NSString *)searchInString:(NSString *)string charStart:(char)start charEnd:(char)end
 {
     int inizio = 0, stop = 0;
+    
     for(int i = 0; i < [string length]; i++) 
     {
         if([string characterAtIndex:i] == start) 
@@ -45,32 +46,15 @@
             break;
         }
     }
-    stop -= inizio;
-    NSString *string2 = [[string substringFromIndex:inizio-1] substringToIndex:stop+1];
     
-    return string2;
+    stop -= inizio;
+    
+    return [[string substringFromIndex:inizio-1] substringToIndex:stop+1];
 }
 
 - (NSString *)searchCharStart:(char)start charEnd:(char)end
 {
-    int inizio = 0, stop = 0;
-    for(int i = 0; i < [self length]; i++) 
-    {
-        if([self characterAtIndex:i] == start) 
-        {
-            inizio = i+1;
-            i += 1;
-        }
-        if([self characterAtIndex:i] == end) 
-        {
-            stop = i;
-            break;
-        }
-    }
-    stop -= inizio;
-    NSString *string = [[self substringFromIndex:inizio-1] substringToIndex:stop+1];
-    
-    return string;
+    return [NSString searchInString:self charStart:start charEnd:end];
 }
 
 - (NSString *)MD5
@@ -135,22 +119,20 @@
 
 - (BOOL)hasString:(NSString *)substring
 {
-    return !([self rangeOfString:substring].location == NSNotFound);
+    return [self hasString:substring caseSensitive:YES];
+}
+
+- (BOOL)hasString:(NSString *)substring caseSensitive:(BOOL)caseSensitive
+{
+    if(caseSensitive)
+        return [self rangeOfString:substring].location != NSNotFound;
+    else
+        return [self.lowercaseString rangeOfString:substring.lowercaseString].location != NSNotFound;
 }
 
 - (BOOL)isEmail
 {
-    NSString *emailRegEx =  
-	@"(?:[a-z0-9!#$%\\&'*+/=?\\^_`{|}~-]+(?:\\.[a-z0-9!#$%\\&'*+/=?\\^_`{|}"  
-	@"~-]+)*|\"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\" 
-	@"x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-"  
-	@"z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:25[0-5"  
-	@"]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-"  
-	@"9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21"  
-	@"-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])";  
-	
-    NSPredicate *regExPredicate = [NSPredicate predicateWithFormat:@"SELF MATCHES %@", emailRegEx];  
-    return [regExPredicate evaluateWithObject:[self lowercaseString]];  
+    return [NSString isEmail:self];
 }
 
 + (BOOL)isEmail:(NSString *)email
@@ -170,43 +152,41 @@
 
 + (NSString *)convertToUTF8Entities:(NSString *)string
 {
-    NSString *isoEncodedString = [[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[
-                        [string stringByReplacingOccurrencesOfString:@"%27" withString:@"'"]
-                                stringByReplacingOccurrencesOfString:[@"%e2%80%99" capitalizedString] withString:@"’"]
-                                stringByReplacingOccurrencesOfString:[@"%2d" capitalizedString] withString:@"-"]
-                                stringByReplacingOccurrencesOfString:[@"%c2%ab" capitalizedString] withString:@"«"]
-                                stringByReplacingOccurrencesOfString:[@"%c2%bb" capitalizedString] withString:@"»"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%80" capitalizedString] withString:@"À"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%82" capitalizedString] withString:@"Â"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%84" capitalizedString] withString:@"Ä"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%86" capitalizedString] withString:@"Æ"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%87" capitalizedString] withString:@"Ç"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%88" capitalizedString] withString:@"È"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%89" capitalizedString] withString:@"É"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%8a" capitalizedString] withString:@"Ê"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%8b" capitalizedString] withString:@"Ë"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%8f" capitalizedString] withString:@"Ï"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%91" capitalizedString] withString:@"Ñ"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%94" capitalizedString] withString:@"Ô"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%96" capitalizedString] withString:@"Ö"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%9b" capitalizedString] withString:@"Û"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%9c" capitalizedString] withString:@"Ü"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a0" capitalizedString] withString:@"à"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a2" capitalizedString] withString:@"â"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a4" capitalizedString] withString:@"ä"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a6" capitalizedString] withString:@"æ"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a7" capitalizedString] withString:@"ç"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a8" capitalizedString] withString:@"è"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%a9" capitalizedString] withString:@"é"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%af" capitalizedString] withString:@"ï"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%b4" capitalizedString] withString:@"ô"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%b6" capitalizedString] withString:@"ö"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%bb" capitalizedString] withString:@"û"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%bc" capitalizedString] withString:@"ü"]
-                                stringByReplacingOccurrencesOfString:[@"%c3%bf" capitalizedString] withString:@"ÿ"]
-                                stringByReplacingOccurrencesOfString:@"%20" withString:@" "];
-    
-    return isoEncodedString;
+    return [[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[string
+                                             stringByReplacingOccurrencesOfString:@"%27" withString:@"'"]
+                                             stringByReplacingOccurrencesOfString:[@"%e2%80%99" capitalizedString] withString:@"’"]
+                                             stringByReplacingOccurrencesOfString:[@"%2d" capitalizedString] withString:@"-"]
+                                             stringByReplacingOccurrencesOfString:[@"%c2%ab" capitalizedString] withString:@"«"]
+                                             stringByReplacingOccurrencesOfString:[@"%c2%bb" capitalizedString] withString:@"»"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%80" capitalizedString] withString:@"À"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%82" capitalizedString] withString:@"Â"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%84" capitalizedString] withString:@"Ä"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%86" capitalizedString] withString:@"Æ"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%87" capitalizedString] withString:@"Ç"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%88" capitalizedString] withString:@"È"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%89" capitalizedString] withString:@"É"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%8a" capitalizedString] withString:@"Ê"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%8b" capitalizedString] withString:@"Ë"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%8f" capitalizedString] withString:@"Ï"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%91" capitalizedString] withString:@"Ñ"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%94" capitalizedString] withString:@"Ô"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%96" capitalizedString] withString:@"Ö"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%9b" capitalizedString] withString:@"Û"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%9c" capitalizedString] withString:@"Ü"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a0" capitalizedString] withString:@"à"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a2" capitalizedString] withString:@"â"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a4" capitalizedString] withString:@"ä"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a6" capitalizedString] withString:@"æ"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a7" capitalizedString] withString:@"ç"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a8" capitalizedString] withString:@"è"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%a9" capitalizedString] withString:@"é"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%af" capitalizedString] withString:@"ï"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%b4" capitalizedString] withString:@"ô"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%b6" capitalizedString] withString:@"ö"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%bb" capitalizedString] withString:@"û"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%bc" capitalizedString] withString:@"ü"]
+                                             stringByReplacingOccurrencesOfString:[@"%c3%bf" capitalizedString] withString:@"ÿ"]
+                                             stringByReplacingOccurrencesOfString:@"%20" withString:@" "];
 }
 
 + (NSString *)encodeToBase64:(NSString *)string
@@ -217,8 +197,7 @@
 
 - (NSString *)encodeToBase64
 {
-    NSData *data = [self dataUsingEncoding:NSUTF8StringEncoding];
-    return [data base64EncodedStringWithOptions:0];
+    return [NSString encodeToBase64:self];
 }
 
 + (NSString *)decodeBase64:(NSString *)string
@@ -229,15 +208,14 @@
 
 - (NSString *)decodeBase64
 {
-    NSData *data = [[NSData alloc] initWithBase64EncodedString:self options:0];
-    return [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
+    return [NSString decodeBase64:self];
 }
 
 - (NSString *)sentenceCapitalizedString
 {
     if(![self length])
     {
-        return [NSString string];
+        return @"";
     }
     NSString *uppercase = [[self substringToIndex:1] uppercaseString];
     NSString *lowercase = [[self substringFromIndex:1] lowercaseString];
@@ -258,28 +236,12 @@
 
 - (NSString *)urlEncode
 {
-    NSMutableString *output = [NSMutableString string];
-    const unsigned char *source = (const unsigned char *)[self UTF8String];
-    int sourceLen = (int)strlen((const char *)source);
-    for(int i = 0; i < sourceLen; ++i)
-    {
-        const unsigned char thisChar = source[i];
-        
-        if(thisChar == ' ')
-        {
-            [output appendString:@"+"];
-        }
-        else if(thisChar == '.' || thisChar == '-' || thisChar == '_' || thisChar == '~' || (thisChar >= 'a' && thisChar <= 'z') || (thisChar >= 'A' && thisChar <= 'Z') || (thisChar >= '0' && thisChar <= '9'))
-        {
-            [output appendFormat:@"%c", thisChar];
-        }
-        else
-        {
-            [output appendFormat:@"%%%02X", thisChar];
-        }
-    }
-    
-    return output;
+    return [self URLEncode];
+}
+
+- (NSString *)URLEncode
+{
+    return [self stringByAddingPercentEncodingWithAllowedCharacters:[NSCharacterSet URLHostAllowedCharacterSet]];
 }
 
 @end