+++ BFKitDemo/BFKitDemo/DetailViewController.m
@@ -530,6 +530,16 @@
             
             break;
         }
+        case DetailTypeBFButton:
+        {
+            self.title = @"BFButton";
+            [_scrollView removeFromSuperview];
+            
+            BFButton *button = [[BFButton alloc] initWithFrame:CGRectMake(20, 84, SCREEN_WIDTH - 40, 50) image:[UIImage imageWithSize:CGSizeMake(SCREEN_WIDTH, 50) backgroundColor:[UIColor colorWithRed:0.301 green:0.550 blue:0.686 alpha:1.000] maskedText:@"BFKit" font:FontNameHelveticaNeueBold fontSize:20] highlightedImage:[UIImage imageWithSize:CGSizeMake(SCREEN_WIDTH, 50) backgroundColor:[UIColor colorWithRed:1.000 green:0.231 blue:0.188 alpha:1.000] maskedText:@"BFKit" font:FontNameHelveticaNeueBold fontSize:20] fadeDuration:1];
+            [self.view addSubview:button];
+            
+            break;
+        }
         case DetailTypeBFLog:
         {
             self.title = @"BFLog";
@@ -538,6 +548,25 @@
             BFLogClear;
             
             BFLog(@"This will be shown only if in DEBUG mode");
+            
+            UITextView *textView = [UITextView initWithFrame:CGRectMake(0, 0, SCREEN_WIDTH, SCREEN_HEIGHT) text:BFLogString color:[UIColor blackColor] font:FontNameHelveticaNeue size:16 alignment:NSTextAlignmentLeft dataDetectorTypes:UIDataDetectorTypeAll editable:NO selectable:NO returnType:UIReturnKeyDefault keyboardType:UIKeyboardTypeDefault secure:NO autoCapitalization:UITextAutocapitalizationTypeNone keyboardAppearance:UIKeyboardAppearanceDefault enablesReturnKeyAutomatically:YES autoCorrectionType:UITextAutocorrectionTypeDefault delegate:nil];
+            [self.view addSubview:textView];
+            
+            break;
+        }
+        case DetailTypeBFPassword:
+        {
+            self.title = @"BFPassword";
+            [_scrollView removeFromSuperview];
+            
+            BFLogClear;
+            
+            NSString *pass1 = @"Password";
+            PasswordStrengthLevel passLevel1 = [BFPassword checkPasswordStrength:pass1];
+            BFLog(@"Password: %@ - Level: %d of %d", pass1, passLevel1, PasswordStrengthLevelVerySecure);
+            NSString *pass2 = @"kqi019ASC.v1|!-2";
+            PasswordStrengthLevel passLevel2 = [BFPassword checkPasswordStrength:pass2];
+            BFLog(@"Password: %@ - Level: %d of %d", pass2, passLevel2, PasswordStrengthLevelVerySecure);
             
             UITextView *textView = [UITextView initWithFrame:CGRectMake(0, 0, SCREEN_WIDTH, SCREEN_HEIGHT) text:BFLogString color:[UIColor blackColor] font:FontNameHelveticaNeue size:16 alignment:NSTextAlignmentLeft dataDetectorTypes:UIDataDetectorTypeAll editable:NO selectable:NO returnType:UIReturnKeyDefault keyboardType:UIKeyboardTypeDefault secure:NO autoCapitalization:UITextAutocapitalizationTypeNone keyboardAppearance:UIKeyboardAppearanceDefault enablesReturnKeyAutomatically:YES autoCorrectionType:UITextAutocorrectionTypeDefault delegate:nil];
             [self.view addSubview:textView];