+++ BFKitDemo/BFKitDemo/UserInterfaceKitViewController.m
@@ -74,7 +74,7 @@
 
 - (NSString *)tableView:(UITableView *)tableView titleForFooterInSection:(NSInteger)section
 {
-	return [NSString stringWithFormat:@"By %@", BFKIT_AUTHOR];
+	return [NSString stringWithFormat:@"By %@ - BFKit v%@ (%@)", BFKIT_AUTHOR, APP_VERSION, APP_BUILD];
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section